+++ ponder/formats.py
@@ -53,8 +53,7 @@ ch = logging.StreamHandler()
 ch.setLevel(logging.DEBUG)
 
 # 创建一个formatter，设置日志格式和颜色
-custom_text = "CreepePonder"
-formatter = CustomColorFormatter(custom_text=custom_text, datefmt='%Y-%m-%d %H:%M:%S')
+formatter = CustomColorFormatter(custom_text="CreepePonder", datefmt='%Y-%m-%d %H:%M:%S')
 ch.setFormatter(formatter)
 
 # 添加handler到logger中
@@ -67,10 +66,14 @@ def get_logger():
 
 """ 指令格式 """
 
-falling_block_cmd = ('summon falling_block {x} {y} {z} {{BlockState:{{Name:"{block_name}",Properties:{{{block_state}}}'
-                     '}},TileEntityData:{nbt},NoGravity:1b,Time:596,Motion:[{motion_x},{motion_y},{motion_z}]}}')
+falling_block_cmd = (
+    'summon falling_block {x} {y} {z} {{BlockState:{{Name:"{block_name}",Properties:{{{block_state}}}'
+    '}},TileEntityData:{nbt},NoGravity:1b,Time:596,Motion:[{motion_x},{motion_y},{motion_z}]}}'
+)
 setblock_cmd = 'setblock {x} {y} {z} {block_name}[{block_state_stripped}]{nbt}'
 summon_cmd = 'summon {entity_name} {x} {y} {z} {nbt}'
-text_display_cmd = ('summon text_display {x} {y} {z} {{text:\'"{text}"\',Tags:["{tag}"],see_through:1b,transformation:'
-                    '{{left_rotation: {rotation}, translation: [0, 0, 0], right_rotation: [0, 0, 0, 1], scale: [1, 1'
-                    ', 1]}}}}')
+text_display_cmd = (
+    'summon text_display {x} {y} {z} {{text:\'"{text}"\',Tags:["{tag}"],see_through:1b,transformation:'
+    '{{left_rotation: {rotation}, translation: [0, 0, 0], right_rotation: [0, 0, 0, 1], scale: [1, 1'
+    ', 1]}}}}'
+)
